+++ head.hpp
@@ -18,24 +18,6 @@ std::map<std::pair<int, int>, Question> dictionary();
 void print_questions_matrix();
 void game();
 
-//// Простая система трекинга пользователя и очков
-//namespace Tracking {
-//    struct User {
-//        std::string name;
-//        int score;
-//    };
-//
-//    //  Инициализация текущего пользователя (если уже инициализирован — перезаписывает)
-//    void init_default_user(const std::string& name);
-//
-//    // Добавление очков текущему пользователю (без выброса исключений)
-//    void add_score(int points) noexcept;
-//
-//    //  Получить текущего пользователя (создаст дефолтного, если не инициализирован)
-//    const User& current_user() noexcept;
-//}
-
-
 void create_user(const std::string& name);
 
 std::pair<char, bool> display_question_with_answers(int row, int col);
